requirements.txt

python-dotenv
azure-identity
azure-ai-projects
azure-ai-agents
semantic-kernel